+++ src/net/sf/persism/TableInfo.java
@@ -1,4 +1,4 @@
 package net.sf.persism;
 
-public final record TableInfo(String name, String schema) {
+record TableInfo(String name, String schema) {
 }