+++ test/net/sf/persism/dao/so/LinkType.java
@@ -1,7 +1,4 @@
 package net.sf.persism.dao.so;
 
 public record LinkType(Integer id, String type) {
-
-    public LinkType {
-    }
 }